requirements.txt
pdfplumber
pandas
openpyxl
camelot-py[cv]